# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API CMS",
        "description": "Todas las rutas de la API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
...
        "/api/v1/login": {
            "post": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Iniciar sesión",
                "operationId": "b860dee2b4dd71501cd036971fc3833a",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "contrasenya"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "pablo@gmail.com"
                                    },
                                    "contrasenya": {
                                        "type": "string",
                                        "example": "1234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token generado"
                    },
                    "400": {
                        "description": "Contraseña incorrecta"
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    }
                }
            }
        },
        "/api/v1/registro": {
            "post": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Registrar un nuevo usuario",
                "operationId": "8025dd8a5875c207988cb09ecbfd0427",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "email",
                                    "contrasenya",
                                    "conf_email"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Pablo"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "pablo@gmail.com"
                                    },
                                    "conf_email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "pablo@gmail.com"
                                    },
                                    "contrasenya": {
                                        "type": "string",
                                        "example": "1234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuario registrado con éxito"
                    },
                    "500": {
                        "description": "Usuario existente, pruebe otro correo"
                    }
                }
            }
        },
        "/api/v1/usuarios": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Obtener una lista de usuarios",
                "operationId": "a25d2373e5467dc335ea0fcc231f571c",
                "responses": {
                    "200": {
                        "description": "Lista de usuarios obtenida con éxito"
                    },
                    "404": {
                        "description": "No se han encontrado usuarios"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/usuarios/{id}": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Obtener detalles de un usuario específico",
                "operationId": "f78387c8739b2423bde11f71ff114d1d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del usuario obtenidos con éxito"
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Actualizar los datos de un usuario",
                "operationId": "a8ab6bb7df25980d2728daeccaa4e645",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "email",
                                    "contrasenya"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Pablo"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "pablo@gmail.com"
                                    },
                                    "contrasenya": {
                                        "type": "string",
                                        "example": "1234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuario actualizado con éxito"
                    },
                    "403": {
                        "description": "No tiene permisos para realizar esta acción"
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Eliminar un usuario",
                "operationId": "25e4dca335cf2c1fec028235d8202147",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuario eliminado con éxito"
                    },
                    "403": {
                        "description": "No tiene permisos para realizar esta acción"
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Usa un token de acceso para autenticar las peticiones",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}